portfolio

+++ src/app/component/prestataire/work-schedule/work-schedule.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { Disponiblite } from 'src/app/interface/disponiblite';
 import { WorkScheduleService } from 'src/app/service/work-schedule.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-work-schedule',
@@ -47,15 +48,55 @@ export class WorkScheduleComponent implements OnInit {
   }
 
   removeDisponibilite(id: number): void {
-    this.disponibiliteService.delete(id)
-      .subscribe(
-        response => {
-          this.refreshList();
-        },
-      );
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
+    });
+  
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.disponibiliteService.delete(id)
+          .subscribe(
+            response => {
+              swalWithBootstrapButtons.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              );
+              this.refreshList();
+            },
+            error => {
+              swalWithBootstrapButtons.fire(
+                'Error!',
+                'An error occurred while deleting the file.',
+                'error'
+              );
+            }
+          );
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        );
+      }
+    });
   }
+  
   disponibiliteForm!: FormGroup;
   onSubmit() {
     // Handle form submission here
   }
+  
 }
